Create README.md
This project is a simple Spring boot application. 
There are two controllers:  Person and Book.
For Simplicity, I have allowed only a Many to one relationship between the above entities. That is A person  can lend as many books as he needs.
The project internally employs H2 in-memory database to aid quicker development.
Run mvn spring-boot:run to see the changes.